working edit model modal
